Add Handgun ammo
