+++ addons/ammunition/config.cpp
@@ -9,16 +9,21 @@ class CfgPatches {
         requiredAddons[] = {
             "tacgt_main",
             "ace_ballistics",
+            "bnae_r1",
             "bnae_RK95",
             "CUP_Weapons_Ammunition",
             "CUP_Weapons_AK12",
             "CUP_Weapons_Bren2",
             "CUP_Weapons_Famas",
+            "CUP_Weapons_M17",
+            "CUP_Weapons_Mk23",
+            "CUP_Weapons_SCAR",
             "CUP_Weapons_X95",
             "hlcweapons_core",
             "hlcweapons_AUG",
             "hlcweapons_SG550",
-            "bnae_r1_virtual"
+            "niaweapons_226",
+            "tacs_weapons"
         };
         author = ECSTRING(main,Authors);
         authors[] = {"Mike"};
