+++ addons/ammunition/handgun_magazines/357.hpp
@@ -1,0 +1,18 @@
+// .357 SIG
+class hlc_12Rnd_357SIG_B_P226;
+class CLASS(12Rnd_357_P226_Magnum): hlc_12Rnd_357SIG_B_P226 {
+    MACRO_SCOPE
+    ammo = QCLASS(357_Magnum);
+    descriptionShort = ".357 Magnum Reload Tracer";
+    displayName = ".357 12Rnd P226 (Magnum)";
+    lastRoundsTracer = 2;
+};
+
+class hlc_8Rnd_357SIG_B_P239;
+class CLASS(8Rnd_357_P239_Magnum): hlc_8Rnd_357SIG_B_P239 {
+    MACRO_SCOPE
+    ammo = QCLASS(357_Magnum);
+    descriptionShort = ".357 Magnum Reload Tracer";
+    displayName = ".357 8Rnd P239 (Magnum)";
+    lastRoundsTracer = 2;
+};
