+++ addons/ammunition/handgun_magazines/45ACP.hpp
@@ -1,0 +1,29 @@
+// .45ACP 1911
+class CUP_7Rnd_45ACP_1911;
+class CLASS(8Rnd_45ACP_1911_Ball): CUP_7Rnd_45ACP_1911 {
+    MACRO_SCOPE
+    ammo = QCLASS(45ACP_Ball);
+    descriptionShort = ".45ACP Ball Reload Tracer";
+    displayName = ".45ACP 8Rnd 1911 (Ball)";
+    lastRoundsTracer = 2;
+};
+
+// .45ACP FNX-45
+class 11Rnd_45ACP_Mag;
+class CLASS(15Rnd_45ACP_FNX45_Ball): 11Rnd_45ACP_Mag {
+    MACRO_SCOPE
+    ammo = QCLASS(45ACP_Ball);
+    descriptionShort = ".45ACP Ball Reload Tracer";
+    displayName = ".45ACP 15Rnd FNX-45 (Ball)";
+    lastRoundsTracer = 2;
+};
+
+// .45ACP MK23 SOCOM
+class CUP_12Rnd_45ACP_mk23;
+class CLASS(12Rnd_45ACP_MK23_Ball): CUP_12Rnd_45ACP_mk23 {
+    MACRO_SCOPE
+    ammo = QCLASS(45ACP_Ball);
+    descriptionShort = ".45ACP Ball Reload Tracer";
+    displayName = ".45ACP 12Rnd MK23 (Ball)";
+    lastRoundsTracer = 2;
+};
