+++ addons/ammunition/handgun_magazines/9x19.hpp
@@ -1,0 +1,78 @@
+class 16Rnd_9x21_Mag;
+class CLASS(17Rnd_9x19mm_Walther_Ball): 16Rnd_9x21_Mag {
+    MACRO_SCOPE
+    ammo = QCLASS(9x19_Ball);
+    descriptionShort = "9x19mm Ball Reload Tracer";
+    displayName = "9mm 17Rnd Walther (Ball)";
+    lastRoundsTracer = 2;
+};
+
+// 9x19 Browning
+class CUP_13Rnd_9x19_Browning_HP;
+class CLASS(13Rnd_9x19_Browning_Ball): CUP_13Rnd_9x19_Browning_HP {
+    MACRO_SCOPE
+    ammo = QCLASS(9x19_Ball);
+    descriptionShort = "9x19mm Ball Reload Tracer";
+    displayName = "9mm 13Rnd Browning (Ball)";
+    lastRoundsTracer = 2;
+};
+
+// 9x19 CZ75
+class CUP_18Rnd_9x19_Phantom;
+class CLASS(18Rnd_9x19_CZ75_Ball): CUP_18Rnd_9x19_Phantom {
+    MACRO_SCOPE
+    ammo = QCLASS(9x19_Ball);
+    descriptionShort = "9x19mm Ball Reload Tracer";
+    displayName = "9mm 18Rnd CZ75 (Ball)";
+    lastRoundsTracer = 2;
+};
+
+// 9x19 Glock
+class CUP_17Rnd_9x19_glock17;
+class CLASS(17Rnd_9x19_Glock_Ball): CUP_17Rnd_9x19_glock17 {
+    MACRO_SCOPE
+    ammo = QCLASS(9x19_Ball);
+    descriptionShort = "9x19mm Ball Reload Tracer";
+    displayName = "9mm 17Rnd Glock (Ball)";
+    lastRoundsTracer = 2;
+};
+
+// 9x19 SIG
+class hlc_15Rnd_9x19_B_P226;
+class CLASS(15Rnd_9x19_P226_Ball): hlc_15Rnd_9x19_B_P226 {
+    MACRO_SCOPE
+    ammo = QCLASS(9x19_Ball);
+    descriptionShort = "9x19mm Ball Reload Tracer";
+    displayName = "9mm 15Rnd P226 (Ball)";
+    lastRoundsTracer = 2;
+};
+
+// 9x19 M17
+class CUP_17Rnd_9x19_M17_Black;
+class CLASS(17Rnd_9x19_M17_Ball): CUP_17Rnd_9x19_M17_Black {
+    MACRO_SCOPE
+    ammo = QCLASS(9x19_Ball);
+    descriptionShort = "9x19mm Ball Reload Tracer";
+    displayName = "9mm 17Rnd M17 (Ball)";
+    lastRoundsTracer = 2;
+};
+
+// 9x19 M9
+class CUP_15Rnd_9x19_M9;
+class CLASS(15Rnd_9x19_M9_Ball): CUP_15Rnd_9x19_M9 {
+    MACRO_SCOPE
+    ammo = QCLASS(9x19_Ball);
+    descriptionShort = "9x19mm Ball Reload Tracer";
+    displayName = "9mm 15Rnd M9 (Ball)";
+    lastRoundsTracer = 2;
+};
+
+// 9x19 SIG P239
+class hlc_10Rnd_9x19_B_P239;
+class CLASS(10Rnd_9x19_P239_Ball): hlc_10Rnd_9x19_B_P239 {
+    MACRO_SCOPE
+    ammo = QCLASS(9x19_Ball);
+    descriptionShort = "9x19mm Ball Reload Tracer";
+    displayName = "9mm 10Rnd P239 (Ball)";
+    lastRoundsTracer = 2;
+};
