+++ addons/ammunition/magwells/handgun.hpp
@@ -1,0 +1,101 @@
+class CBA_9x19_HiPower {
+    ADDON[] = {
+        QCLASS(13Rnd_9x19_Browning_Ball)
+    };
+};
+
+class CBA_9x19_CZP09 {
+    ADDON[] = {
+        QCLASS(17Rnd_9x19mm_Walther_Ball)
+    };
+};
+
+class CBA_9x19_CZ75_FULL {
+    ADDON[] = {
+        QCLASS(18Rnd_9x19_CZ75_Ball)
+    };
+};
+
+class CBA_9x19_Walther {
+    ADDON[] = {
+        QCLASS(17Rnd_9x19mm_Walther_Ball)
+    };
+};
+
+class CBA_9x19_Glock_Full {
+    ADDON[] = {
+        QCLASS(17Rnd_9x19_Glock_Ball)
+    };
+};
+
+class CBA_9x19_P226 {
+    ADDON[] = {
+        QCLASS(15Rnd_9x19_P226_Ball)
+    };
+};
+
+class CBA_9x19_P228 {
+    ADDON[] = {
+        QCLASS(15Rnd_9x19_P226_Ball)
+    };
+};
+
+class CBA_9x19_M17 {
+    ADDON[] = {
+        QCLASS(17Rnd_9x19_M17_Ball)
+    };
+};
+
+class CBA_9x19_M9 {
+    ADDON[] = {
+        QCLASS(15Rnd_9x19_M9_Ball)
+    };
+};
+
+class CBA_9x19_P239 {
+    ADDON[] = {
+        QCLASS(10Rnd_9x19_P239_Ball)
+    };
+};
+
+class CBA_45ACP_FNX45 {
+    ADDON[] = {
+        QCLASS(15Rnd_45ACP_FNX45_Ball)
+    };
+};
+
+class CBA_45ACP_MK23 {
+    ADDON[] = {
+        QCLASS(12Rnd_45ACP_MK23_Ball)
+    };
+};
+
+class CBA_45ACP_1911 {
+    ADDON[] = {
+        QCLASS(8Rnd_45ACP_1911_Ball)
+    };
+};
+
+class CBA_357SIG_P239 {
+    ADDON[] = {
+        QCLASS(8Rnd_357_P239_Magnum)
+    };
+};
+
+class CBA_357SIG_P226 {
+    ADDON[] = {
+        QCLASS(12Rnd_357_P226_Magnum)
+    };
+};
+
+class CLASS(Cylinder_357) {
+    ADDON[] = {
+        QCLASS(6Rnd_357)
+    };
+};
+
+class CLASS(Cylinder_44) {
+    ADDON[] = {
+        QCLASS(6Rnd_44)
+    };
+};
