+++ addons/ammunition/magwells/rifle.hpp
@@ -1,0 +1,456 @@
+class CBA_12g_2rnds {
+    ADDON[] = {
+        QCLASS(2Rnd_P_000)
+    };
+};
+
+class HunterShotgun_01_12GA {
+    ADDON[] = {
+        QCLASS(2Rnd_P_000)
+    };
+};
+
+class CBA_12g_8rnds {
+    ADDON[] = {
+        QCLASS(8Rnd_P_000),
+        QCLASS(8Rnd_S_AP20)
+    };
+};
+
+class CBA_12g_6rnds {
+    ADDON[] = {
+        QCLASS(6Rnd_P_000),
+        QCLASS(6Rnd_S_AP20)
+    };
+};
+
+class UBS_12GA {
+    ADDON[] = {
+        QCLASS(6Rnd_P_UGL),
+        QCLASS(6Rnd_S_UGL)
+    };
+};
+
+class CBA_12g_SAIGA {
+    ADDON[] = {
+        QCLASS(8Rnd_Saiga_000),
+        QCLASS(8Rnd_Saiga_AP20),
+        QCLASS(20Rnd_Saiga_000),
+        QCLASS(20Rnd_Saiga_AP20)
+    };
+};
+
+class CBA_9x19_MP5 {
+    ADDON[] = {
+        QCLASS(30Rnd_9x19_Ball),
+        QCLASS(30Rnd_9x19_Ball_Tracer),
+        QCLASS(30Rnd_9x19_EPR),
+        QCLASS(30Rnd_9x19_AP)
+    };
+};
+
+class CBA_9x39_VSS {
+    ADDON[] = {
+        QCLASS(30Rnd_9x39_Ball),
+        QCLASS(30Rnd_9x39_Ball_Tracer),
+        QCLASS(30Rnd_9x39_EPR),
+        QCLASS(30Rnd_9x39_AP)
+    };
+};
+
+class CBA_300BLK_STANAG {
+    ADDON[] = {
+        QCLASS(30Rnd_300AAC_Ball),
+        QCLASS(30Rnd_300AAC_Ball_Tracer),
+        QCLASS(30Rnd_300AAC_EPR),
+        QCLASS(30Rnd_300AAC_AP)
+    };
+};
+
+class CLASS(9x19_MP5) { // Prevents regular BI Magazines fitting the CUP MP5s as no model shows.
+    ADDON[] = {
+        QCLASS(30Rnd_9x19_Ball_MP5),
+        QCLASS(30Rnd_9x19_Ball_Tracer_MP5),
+        QCLASS(30Rnd_9x19_EPR_MP5),
+        QCLASS(30Rnd_9x19_AP_MP5)
+    };
+};
+
+class CBA_9x19_ScorpionEvo3 {
+    ADDON[] = {
+        QCLASS(30Rnd_9x19_Ball),
+        QCLASS(30Rnd_9x19_Ball_Tracer),
+        QCLASS(30Rnd_9x19_EPR),
+        QCLASS(30Rnd_9x19_AP)
+    };
+};
+
+class CBA_9x19_PP19 {
+    ADDON[] = {
+        QCLASS(64Rnd_9x19_Ball_Bizon),
+        QCLASS(64Rnd_9x19_Ball_Tracer_Bizon),
+        QCLASS(64Rnd_9x19_EPR_Bizon),
+        QCLASS(64Rnd_9x19_AP_Bizon)
+    };
+};
+
+class CBA_9x19_Vityaz {
+    ADDON[] = {
+        QCLASS(30Rnd_9x19_Ball_Vityaz),
+        QCLASS(30Rnd_9x19_Ball_Tracer_Vityaz),
+        QCLASS(30Rnd_9x19_EPR_Vityaz),
+        QCLASS(30Rnd_9x19_AP_Vityaz)
+    };
+};
+
+class CBA_46x30_MP7 {
+    ADDON[] = {
+        QCLASS(40Rnd_46x30_Ball),
+        QCLASS(40Rnd_46x30_Ball_Tracer),
+        QCLASS(40Rnd_46x30_EPR),
+        QCLASS(40Rnd_46x30_AP)
+    };
+};
+
+class CBA_545x39_AK {
+    ADDON[] = {
+        QCLASS(30Rnd_545x39_Ball),
+        QCLASS(30Rnd_545x39_Ball_Black),
+        QCLASS(30Rnd_545x39_Ball_Tracer),
+        QCLASS(30Rnd_545x39_Ball_Tracer_Black),
+        QCLASS(30Rnd_545x39_EPR),
+        QCLASS(30Rnd_545x39_EPR_Black),
+        QCLASS(30Rnd_545x39_AP),
+        QCLASS(30Rnd_545x39_AP_Black),
+        QCLASS(60Rnd_545x39_Ball_Quadstack),
+        QCLASS(60Rnd_545x39_Ball_Tracer_Quadstack),
+        QCLASS(60Rnd_545x39_EPR_Quadstack)
+    };
+};
+
+class CBA_545x39_Fort {
+    ADDON[] = {
+        QCLASS(30Rnd_545x39_Ball_Fort),
+        QCLASS(30Rnd_545x39_Ball_Tracer_Fort),
+        QCLASS(30Rnd_545x39_EPR_Fort),
+        QCLASS(30Rnd_545x39_AP_Fort)
+    };
+};
+
+class CBA_45ACP_Glock_Full {
+    ADDON[] = {
+        QCLASS(25Rnd_45ACP_Ball),
+        QCLASS(25Rnd_45ACP_Ball_Tracer),
+        QCLASS(25Rnd_45ACP_EPR)
+    };
+};
+
+class CBA_57x28_P90 {
+    ADDON[] = {
+        QCLASS(50Rnd_57x28_Ball),
+        QCLASS(50Rnd_57x28_Ball_Tracer),
+        QCLASS(50Rnd_57x28_EPR),
+        QCLASS(50Rnd_57x28_AP)
+    };
+};
+
+class CBA_580x42_TYPE95 {
+    ADDON[] = {
+        QCLASS(30Rnd_58x42_Ball),
+        QCLASS(30Rnd_58x42_Ball_Tracer),
+        QCLASS(30Rnd_58x42_EPR),
+        QCLASS(30Rnd_58x42_AP)
+    };
+};
+
+class CBA_580x42_TYPE95_XL {
+    ADDON[] = {
+        QCLASS(100Rnd_58x42_Ball),
+        QCLASS(100Rnd_58x42_Ball_Tracer),
+        QCLASS(100Rnd_58x42_EPR)
+    };
+};
+
+class CBA_556x45_AK {
+    ADDON[] = {
+        QCLASS(30Rnd_556x45_Ball_AK),
+        QCLASS(30Rnd_556x45_Ball_Tracer_AK),
+        QCLASS(30Rnd_556x45_EPR_AK),
+        QCLASS(30Rnd_556x45_AP_AK)
+    };
+};
+
+class CBA_556x45_G36 {
+    ADDON[] = {
+        QCLASS(30Rnd_556x45_Ball_G36),
+        QCLASS(30Rnd_556x45_Ball_Tracer_G36),
+        QCLASS(30Rnd_556x45_EPR_G36),
+        CLASS(30Rnd_556x45_AP_G36),
+        QCLASS(100Rnd_556x45_Ball_BetaC),
+        QCLASS(100Rnd_556x45_Ball_Tracer_BetaC),
+        QCLASS(100Rnd_556x45_EPR_BetaC)
+    };
+};
+
+class CBA_556x45_FAMAS {
+    ADDON[] = {
+        QCLASS(25Rnd_556x45_Ball_FAMAS),
+        QCLASS(25Rnd_556x45_Ball_Tracer_FAMAS),
+        QCLASS(25Rnd_556x45_EPR_FAMAS),
+        QCLASS(25Rnd_556x45_AP_FAMAS)
+    };
+};
+
+class CBA_556x45_SG550 {
+    ADDON[] = {
+        QCLASS(30Rnd_556x45_Ball_SG),
+        QCLASS(30Rnd_556x45_Ball_Tracer_SG),
+        QCLASS(30Rnd_556x45_EPR_SG),
+        QCLASS(30Rnd_556x45_AP_SG)
+    };
+};
+
+class CBA_556x45_STANAG {
+    ADDON[] = {
+        QCLASS(30Rnd_556x45_Ball_EMAG),
+        QCLASS(30Rnd_556x45_Ball_Tracer_EMAG),
+        QCLASS(30Rnd_556x45_EPR_EMAG),
+        QCLASS(30Rnd_556x45_AP_EMAG),
+        QCLASS(30Rnd_556x45_Ball_EMAG_Tan),
+        QCLASS(30Rnd_556x45_Ball_Tracer_EMAG_Tan),
+        QCLASS(30Rnd_556x45_EPR_EMAG_Tan),
+        QCLASS(30Rnd_556x45_AP_EMAG_Tan),
+        QCLASS(30Rnd_556x45_Ball_PMAG),
+        QCLASS(30Rnd_556x45_Ball_Tracer_PMAG),
+        QCLASS(30Rnd_556x45_EPR_PMAG),
+        QCLASS(30Rnd_556x45_AP_PMAG),
+        QCLASS(30Rnd_556x45_Ball_PMAG_Tan),
+        QCLASS(30Rnd_556x45_Ball_Tracer_PMAG_Tan),
+        QCLASS(30Rnd_556x45_EPR_PMAG_Tan),
+        QCLASS(30Rnd_556x45_AP_PMAG_Tan),
+        QCLASS(60Rnd_556x45_Ball_Surefire),
+        QCLASS(60Rnd_556x45_Ball_Tracer_Surefire),
+        QCLASS(60Rnd_556x45_EPR_Surefire),
+        QCLASS(100Rnd_556x45_Ball_BetaC),
+        QCLASS(100Rnd_556x45_Ball_Tracer_BetaC),
+        QCLASS(100Rnd_556x45_EPR_BetaC)
+    };
+};
+
+class CBA_556x45_STANAG_2D_XL {
+    ADDON[] = {
+        QCLASS(150Rnd_556x45_Ball_Drum),
+        QCLASS(150Rnd_556x45_Ball_Tracer_Drum),
+        QCLASS(150Rnd_556x45_EPR_Drum)
+    };
+};
+
+class CBA_556x45_MINIMI {
+    ADDON[] = {
+        QCLASS(200Rnd_556x45_Ball_Belt),
+        QCLASS(200Rnd_556x45_Ball_Tracer_Belt),
+        QCLASS(200Rnd_556x45_EPR_Belt)
+    };
+};
+
+class CBA_556x45_STEYR {
+    ADDON[] = {
+        QCLASS(30Rnd_556x45_Ball_AUG),
+        QCLASS(30Rnd_556x45_Ball_Tracer_AUG),
+        QCLASS(30Rnd_556x45_EPR_AUG),
+        QCLASS(30Rnd_556x45_AP_AUG),
+        QCLASS(42Rnd_556x45_Ball_AUG),
+        QCLASS(42Rnd_556x45_Ball_Tracer_AUG),
+        QCLASS(42Rnd_556x45_EPR_AUG),
+        QCLASS(42Rnd_556x45_AP_AUG)
+    };
+};
+
+class CBA_65x39_MSBS {
+    ADDON[] = {
+        QCLASS(30Rnd_65x39_Ball_MSBS),
+        QCLASS(30Rnd_65x39_Ball_Tracer_MSBS),
+        QCLASS(30Rnd_65x39_EPR_MSBS),
+        QCLASS(30Rnd_65x39_AP_MSBS)
+    };
+};
+
+class CBA_65x39_MX {
+    ADDON[] = {
+        QCLASS(30Rnd_65x39_Ball_MX_Sand),
+        QCLASS(30Rnd_65x39_Ball_Tracer_MX_Sand),
+        QCLASS(30Rnd_65x39_EPR_MX_Sand),
+        QCLASS(30Rnd_65x39_AP_MX_Sand),
+        QCLASS(30Rnd_65x39_Ball_MX_Black),
+        QCLASS(30Rnd_65x39_Ball_Tracer_MX_Black),
+        QCLASS(30Rnd_65x39_EPR_MX_Black),
+        QCLASS(30Rnd_65x39_AP_MX_Black)
+    };
+};
+
+class CBA_65x39_MX_XL {
+    ADDON[] = {
+        QCLASS(100Rnd_65x39_Ball_MX_LSW_Sand),
+        QCLASS(100Rnd_65x39_Ball_Tracer_MX_LSW_Sand),
+        QCLASS(100Rnd_65x39_EPR_MX_LSW_Sand),
+        QCLASS(100Rnd_65x39_Ball_MX_LSW_Black),
+        QCLASS(100Rnd_65x39_Ball_Tracer_MX_LSW_Black),
+        QCLASS(100Rnd_65x39_EPR_MX_LSW_Black)
+    };
+};
+
+class CBA_65x39_Katiba {
+    ADDON[] = {
+        QCLASS(30Rnd_65x39_Ball_Katiba),
+        QCLASS(30Rnd_65x39_Ball_Tracer_Katiba),
+        QCLASS(30Rnd_65x39_EPR_Katiba),
+        QCLASS(30Rnd_65x39_AP_Katiba)
+    };
+};
+
+class CBA_65x39_Mk200 {
+    ADDON[] = {
+        QCLASS(200Rnd_65x39_Ball_Belt),
+        QCLASS(200Rnd_65x39_Ball_Tracer_Belt),
+        QCLASS(200Rnd_65x39_EPR_Belt)
+    };
+};
+
+class CBA_65x39_QBU {
+    ADDON[] = {
+        QCLASS(20Rnd_65x39_Ball),
+        QCLASS(20Rnd_65x39_Ball_Tracer),
+        QCLASS(20Rnd_65x39_AP)
+    };
+};
+
+class CBA_68SPC_STANAG {
+    ADDON[] = {
+        QCLASS(30Rnd_68x43_Ball_ACR),
+        QCLASS(30Rnd_68x43_Ball_Tracer_ACR),
+        QCLASS(30Rnd_68x43_EPR_ACR),
+        QCLASS(30Rnd_68x43_AP_ACR)
+    };
+};
+
+class CBA_762x39_AK {
+    ADDON[] = {
+        QCLASS(30Rnd_762x39_Ball_Plastic_Black),
+        QCLASS(30Rnd_762x39_Ball_Tracer_Plastic_Black),
+        QCLASS(30Rnd_762x39_EPR_Plastic_Black),
+        QCLASS(30Rnd_762x39_AP_Plastic_Black),
+        QCLASS(30Rnd_762x39_Ball_Plastic_Tan),
+        QCLASS(30Rnd_762x39_Ball_Tracer_Plastic_Tan),
+        QCLASS(30Rnd_762x39_EPR_Plastic_Tan),
+        QCLASS(30Rnd_762x39_AP_Plastic_Tan),
+        QCLASS(30Rnd_762x39_Ball_Metal),
+        QCLASS(30Rnd_762x39_Ball_Tracer_Metal),
+        QCLASS(30Rnd_762x39_EPR_Metal),
+        QCLASS(30Rnd_762x39_AP_Metal)
+    };
+};
+
+class CBA_762x39_RPK {
+    ADDON[] = {
+        QCLASS(75Rnd_762x39_Ball_Drum),
+        QCLASS(75Rnd_762x39_Ball_Tracer_Drum),
+        QCLASS(75Rnd_762x39_EPR_Drum)
+    };
+};
+
+class CLASS(762x39_CZ) {
+    ADDON[] = {
+        QCLASS(30Rnd_762x39_Ball_CZ),
+        QCLASS(30Rnd_762x39_Ball_Tracer_CZ),
+        QCLASS(30Rnd_762x39_EPR_CZ),
+        QCLASS(30Rnd_762x39_AP_CZ)
+    };
+};
+
+class CBA_762x51_AR10 {
+    ADDON[] = {
+        QCLASS(20Rnd_762x51_Ball),
+        QCLASS(20Rnd_762x51_Ball_Tracer),
+        QCLASS(20Rnd_762x51_EPR),
+        QCLASS(20Rnd_762x51_AP)
+    };
+};
+
+class CBA_762x51_FAL {
+    ADDON[] = {
+        QCLASS(20Rnd_762x51_Ball_FAL),
+        QCLASS(20Rnd_762x51_Ball_Tracer_FAL),
+        QCLASS(20Rnd_762x51_EPR_FAL),
+        QCLASS(20Rnd_762x51_AP_FAL)
+    };
+};
+
+class CBA_762x51_G3 {
+    ADDON[] = {
+        QCLASS(20Rnd_762x51_Ball),
+        QCLASS(20Rnd_762x51_Ball_Tracer),
+        QCLASS(20Rnd_762x51_EPR),
+        QCLASS(20Rnd_762x51_AP)
+    };
+};
+
+class CBA_762x51_SCAR {
+    ADDON[] = {
+        QCLASS(20Rnd_762x51_Ball_SCAR),
+        QCLASS(20Rnd_762x51_Ball_Tracer_SCAR),
+        QCLASS(20Rnd_762x51_EPR_SCAR),
+        QCLASS(20Rnd_762x51_AP_SCAR)
+    };
+};
+
+class CBA_762x51_HK417 {
+    ADDON[] = {
+        QCLASS(20Rnd_762x51_Ball),
+        QCLASS(20Rnd_762x51_Ball_Tracer),
+        QCLASS(20Rnd_762x51_EPR),
+        QCLASS(20Rnd_762x51_AP)
+    };
+};
+
+class CBA_762x51_M14 {
+    ADDON[] = {
+        QCLASS(20Rnd_762x51_Ball),
+        QCLASS(20Rnd_762x51_Ball_Tracer),
+        QCLASS(20Rnd_762x51_EPR),
+        QCLASS(20Rnd_762x51_AP)
+    };
+};
+
+class CBA_762x51_SR25 {
+    ADDON[] = {
+        QCLASS(20Rnd_762x51_Ball),
+        QCLASS(20Rnd_762x51_Ball_Tracer),
+        QCLASS(20Rnd_762x51_EPR),
+        QCLASS(20Rnd_762x51_AP)
+    };
+};
+
+class CBA_762x54R_SVD {
+    ADDON[] = {
+        QCLASS(10Rnd_762x54r_Ball),
+        CLASS(10Rnd_762x54r_Ball_Tracer),
+        QCLASS(10Rnd_762x54r_EPR),
+        QCLASS(10Rnd_762x54r_AP)
+    };
+};
+
+class CBA_762x51_LINKS {
+    ADDON[] = {
+        QCLASS(100Rnd_762x51_Ball_Belt),
+        QCLASS(100Rnd_762x51_Ball_Tracer_Belt),
+        CLASS(100Rnd_762x51_EPR_Belt)
+    };
+};
+
+class CBA_762x54R_LINKS {
+    ADDON[] = {
+        QCLASS(150Rnd_762x54r_Ball_Belt),
+        QCLASS(150Rnd_762x54r_Ball_Tracer_Belt),
+        QCLASS(150Rnd_762x54r_EPR_Belt)
+    };
+};
