--- addons/chiappa/CfgMagazineWells.hpp
@@ -1,13 +1,0 @@
-class CfgMagazineWells {
-    class CLASS(Cylinder_357) {
-        ADDON[] = {
-            QCLASS(6Rnd_357)
-        };
-    };
-
-    class CLASS(Cylinder_44) {
-        ADDON[] = {
-            QCLASS(6Rnd_44)
-        };
-    };
-};
